\name{bp.polys}
\docType{data}
\alias{bp.polys}
\title{Substock Polygons}
\description{
This is a list of two polygons representing the extent of the two substocks (= breeding populations) in the TOSSM
examples.
}
\usage{bp.polys}
\format{list (length = 2) of gpclib polygon objects (\code{\link[<gpclib>:<gpclib>-package]{<gpclib>}})}
\keyword{datasets}