\name{example.historic.polys}
\docType{data}
\alias{example.historic.polys}
\title{Historic Polygons}
\description{
This is a list of two polygons representing the extent of the two areas of historic harvest in the second 
\code{run.tossm} example.

To create your own historic polygons, see \code{\link{example.bp.polys}} 
}
\usage{example.historic.polys}
\format{list (length = 2) of gpclib polygon objects (\code{\link[<gpclib>:<gpclib>-package]{<gpclib>}})}
\keyword{datasets}