\name{example.landscape}
\docType{data}
\alias{example.landscape}
\title{Rmetasim Landscape for TOSSM Examples}
\description{
This is an example Rmetasim dataset which serves as a starting population for a TOSSM simulation.  This particular 
dataset was chosen in part because it is one of the smaller-sized TOSSM datasets.  There are two substocks, each with
a carrying capacity of 2500 individuals.  The dispersal rate between the two stocks is 5X10-5. \cr
This dataset is of the Archetype II variety, scenario 63. \cr
Visit \url{http://swfscdata.nmfs.noaa.gov/TOSSM/} for more information, as well as to download other TOSSM datasets.
}
\usage{example.landscape}
\format{Rmetasim landscape object (\code{\link[<rmetasim>:<rmetasim>-package]{<rmetasim>}})}
\keyword{datasets}