\name{example.sample.polys}
\docType{data}
\alias{example.sample.polys}
\title{Sampling Polygons}
\description{
This is a list of eight polygons representing the extent of the sampling areas in the TOSSM examples.

To create your own sampling polygons, see \code{\link{example.bp.polys}} 
}
\usage{example.sample.polys}
\format{list (length = 8) of gpclib polygon objects (\code{\link[<gpclib>:<gpclib>-package]{<gpclib>}})}
\keyword{datasets}