\name{historic.polys}
\docType{data}
\alias{historic.polys}
\title{Historic Polygons}
\description{
This is a list of two polygons representing the extent of the two areas of historic harvest in the 2nd 
\code{run.tossm} example.
}
\usage{historic.polys}
\format{list (length = 2) of gpclib polygon objects (\code{\link[<gpclib>:<gpclib>-package]{<gpclib>}})}
\keyword{datasets}