\name{sample.polys}
\docType{data}
\alias{sample.polys}
\title{Sampling Polygons}
\description{
This is a list of eight polygons representing the extent of the sampling areas in the TOSSM examples.
}
\usage{sample.polys}
\format{list (length = 8) of gpclib polygon objects (\code{\link[<gpclib>:<gpclib>-package]{<gpclib>}})}
\keyword{datasets}