\name{tossm-internal}
\alias{abind.mvb}
\alias{adjust.pop.dyn}
\alias{agg.gs.series}
\alias{agg.gs.to.gtypes}
\alias{as.smat}
\alias{breed.grow.die}
\alias{calc.effort}
\alias{CLA}
\alias{count.alleles.by.f}
\alias{count.whales.b}
\alias{def.depleter}
\alias{def.set.coords}
\alias{divvy.BPs}
\alias{divvy.catch}
\alias{fast.trans}
\alias{fixed.BSA}
\alias{generate.coords}
\alias{get.FIMA}
\alias{get.n.in.ss}
\alias{get.SRM}
\alias{get.voronoi}
\alias{gpcpoly.centroid}
\alias{group.membership}
\alias{harvest}
\alias{humpbackism}
\alias{hyptest.seq.BSA}
\alias{i.file}
\alias{in.gpc}
\alias{index.to.state}
\alias{km.simple.demo}
\alias{line.intersect}
\alias{mixmat}
\alias{old.bgd}
\alias{original!object!list}
\alias{PBR}
\alias{poly.intersect}
\alias{poly.union}
\alias{prepare.rland}
\alias{print.printinfo}
\alias{print.tossm.obj}
\alias{randomism}
\alias{raw.CLA}
\alias{rsample}
\alias{run.mvb.example2}
\alias{set.interval.polys}
\alias{sim.abund.est}
\alias{single.MU.BSA}
\alias{sqr}
\alias{structure.batch.BSA}
\alias{TAC.split.by.f}
\alias{tossm.diagnostics}
\alias{tossm.plotter}
\alias{tossm.round}
\alias{trajectory.plotter}
\alias{write.mainparams}
\title{Internal tossm functions}
\description{Internal tossm functions, not to be called directly by user}
\keyword{internal}